+++ nextstep-di/src/main/java/nextstep/di/factory/beans/BeanRecipe.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.factory.beans;
 
 public interface BeanRecipe {
     Object bakeBean(Object... params);
