+++ nextstep-di/src/main/java/nextstep/di/factory/context/BeanFactoryUtils.java
@@ -1,0 +1,31 @@
+package nextstep.di.factory.context;
+
+import com.google.common.collect.Sets;
+
+import java.util.Optional;
+import java.util.Set;
+
+public class BeanFactoryUtils {
+    /**
+     * 인자로 전달되는 클래스의 구현 클래스. 만약 인자로 전달되는 Class가 인터페이스가 아니면 전달되는 인자가 구현 클래스,
+     * 인터페이스인 경우 BeanFactory가 관리하는 모든 클래스 중에 인터페이스를 구현하는 클래스를 찾아 반환
+     *
+     * @param injectedClazz
+     * @param preInstantiateBeans
+     * @return
+     */
+    public static Optional<Class<?>> findConcreteClass(Class<?> injectedClazz, Set<Class<?>> preInstantiateBeans) {
+        if (!injectedClazz.isInterface()) {
+            return Optional.of(injectedClazz);
+        }
+
+        for (Class<?> clazz : preInstantiateBeans) {
+            Set<Class<?>> interfaces = Sets.newHashSet(clazz.getInterfaces());
+            if (interfaces.contains(injectedClazz)) {
+                return Optional.of(clazz);
+            }
+        }
+
+        return Optional.empty();
+    }
+}
