+++ nextstep-di/src/test/java/nextstep/di/factory/BeanFactoryErrorTest.java
@@ -1,5 +1,8 @@
 package nextstep.di.factory;
 
+import nextstep.di.factory.beans.BeanRecipe;
+import nextstep.di.factory.beans.ConstructorBeanRecipe;
+import nextstep.di.factory.context.BeanFactory;
 import nextstep.di.factory.error.BadBean;
 import nextstep.di.factory.error.Circular;
 import nextstep.di.factory.error.GoodBean;
